:sparkles: Add Get function from mongodb to services page
