+++ src/pages/Services/ElektrikerList/ElektrikerList.module.css
@@ -1,8 +1,15 @@
 .main {
-  background-color: var(--color-primary);
+  background-color: var(--color-tertiary);
 }
 
 .container {
+  display: grid;
+  grid-template-rows: auto;
+  row-gap: 1rem;
   background-color: var(--color-tertiary);
-  margin: 100px;
+}
+
+.row {
+  background: var(--color-secondary);
+  margin-right: 10px;
 }
